+++ src/utils.c
@@ -6,7 +6,7 @@
 /*   By: emurillo <emurillo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/31 16:09:35 by emurillo          #+#    #+#             */
-/*   Updated: 2025/06/07 12:08:06 by emurillo         ###   ########.fr       */
+/*   Updated: 2025/06/10 15:51:46 by emurillo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,12 @@
 void	init_philo(t_args *args, char **av, int ac)
 {
 	args->num_of_phil = ft_atoi(av[1]);
-	args->time_to_die = (size_t)ft_atoi(av[2]);
+	args->time_to_die = ft_atoi(av[2]);
 
-	args->time_to_eat = (size_t)ft_atoi(av[3]);
-	args->time_to_sleep = (size_t)ft_atoi(av[4]);
+	args->time_to_eat = ft_atoi(av[3]);
+	args->time_to_sleep = ft_atoi(av[4]);
 	if (ac == 6)
-		args->meals_to_have = (size_t)ft_atoi(av[5]);
+		args->meals_to_have = ft_atoi(av[5]);
 	if (args->time_to_eat
 		&& args->time_to_die && args->time_to_sleep)
 		printf("struct init correctly\n");
@@ -31,12 +31,12 @@ void	print_struct(t_args *args, int ac)
 	printf("Struct contains the next values.\n");
 	if (ac == 6)
 	{
-		printf("N: [%d], die: [%ld] eat:[%ld], sleep[%ld], meals: [%ld]\n",
+		printf("N: [%d], die: [%d] eat:[%d], sleep[%d], meals: [%d]\n",
 			args->num_of_phil, args->time_to_die, args->time_to_eat, \
 			args->time_to_sleep, args->meals_to_have);
 	}
 	else
-		printf("N: [%d], die: [%ld] eat:[%ld], sleep[%ld]\n",
+		printf("N: [%d], die: [%d] eat:[%d], sleep[%d]\n",
 			args->num_of_phil, args->time_to_die, args->time_to_eat, \
 			args->time_to_sleep);
 }
